requirements.txt
Django>=4.2
djangorestframework
drf-yasg
gunicorn
psycopg2-binary
python-dotenv